Switch to 32-bit mode

+++ leb128.h
@@ -5,19 +5,19 @@
 
 
 typedef struct {
-  uint64_t value;
+  uint32_t value;
   uint8_t size;
 } value_size;
 
 
-uint8_t uleb128_bytes(uint64_t value);
+uint8_t uleb128_bytes(uint32_t value);
 
-uint8_t uleb128_encode(uint64_t value, uint8_t *data);
+uint8_t uleb128_encode(uint32_t value, uint8_t *data);
 
 value_size uleb128_decode(uint8_t *data);
 
-uint8_t leb128_bytes(int64_t value);
+uint8_t leb128_bytes(int32_t value);
 
-uint8_t leb128_encode(int64_t value, uint8_t *data);
+uint8_t leb128_encode(int32_t value, uint8_t *data);
 
 value_size leb128_decode(uint8_t *data);
